backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
requests
python-dotenv
openai
langchain
anthropic
langchain-google-genai
langchain-community
pyyaml